+++ __tests__/Unit/Components/IdleUsers/Card.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import Card from '@/components/idleUsers/card';
-import { calculateIdleSince } from '@/components/idleUsers/card';
+import getIdleSinceText from '@/helperFunctions/getIdleSinceText';
 
 const user = {
     id: "H3vNvHtFfp1Y57tPNoQ1",
@@ -23,7 +23,7 @@ describe("Idle User Card", () => {
         render(<Card user={user} />)
         const userImage = screen.getByTestId('user-image');
         const fullName = screen.getByText(user.full_name);
-        const idleSinceText = calculateIdleSince(user.currentStatus.from);
+        const idleSinceText = getIdleSinceText(user.currentStatus.from);
         const idleSince = screen.getByTestId('idle-since');
         expect(idleSince).toHaveTextContent(idleSinceText);
         expect(fullName).toBeInTheDocument();
