+++ src/components/constants/date.js
@@ -1,3 +1,5 @@
 export const THOUSAND_MILLI_SECONDS = 1000;
 export const FOURTEEN_DAYS = 14;
 export const SECONDS_IN_A_DAY = 86400;
+export const TOTAL_MILLISECONDS_IN_A_DAY = THOUSAND_MILLI_SECONDS * 24;
+export const TOTAL_MILLISECONDS_IN_A_HOUR = THOUSAND_MILLI_SECONDS * 60 * 60
